taxdiv: Taxonomic Diversity Indices Using Deng Entropy

Calculates taxonomic diversity indices for ecological community data using Deng entropy framework and classical approaches (Shannon, Simpson, Clarke & Warwick). Provides functions for computing taxonomic distinctness, average taxonomic distinctness (AvTD/Delta+), variation in taxonomic distinctness (VarTD/Lambda+), and Deng entropy-based measures that incorporate taxonomic hierarchy information. Includes tools for constructing taxonomic trees and computing pairwise taxonomic distances.

See also

Author

Maintainer: Muhammet Murat Gorgoz muratgorgoz350@gmail.com (ORCID)

Authors:

Mehmet Guvenc Negiz

Kursad Ozkan